Add files via upload
This is billing software which can be used in real life and solved the  problem of making basic bill for the small & medium business, and also this whole program i'd using completely c language. If there is a difficulty in any aspect of  this code, we will try to solve your problem. Thank you...